requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.5.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-slugify==8.0.1
pydantic[email]==2.7.1
python-dotenv==1.0.1 